SAK-43102 - Extract AnnouncementWrapper
This is a first step toward cleaning up AnnouncementAction. It extracts
the AnnouncementWrapper nested class, tidies up the handling of a couple
of entangled constants, removes a few references to TimeService, and
adds a couple of minimal tests to get things moving.
