+++ announcement/announcement-tool/tool/src/java/org/sakaiproject/announcement/tool/AnnouncementAction.java
@@ -21,6 +21,8 @@
 
 package org.sakaiproject.announcement.tool;
 
+import java.time.Duration;
+import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
@@ -32,26 +34,21 @@ import java.util.Map;
 import java.util.HashMap;
 import java.util.NoSuchElementException;
 import java.util.Properties;
-import java.util.Stack;
 import java.util.Vector;
 
 import lombok.extern.slf4j.Slf4j;
 
 import org.apache.commons.lang3.StringUtils;
-
-import org.w3c.dom.Document;
-import org.w3c.dom.Element;
-
 import org.sakaiproject.announcement.api.AnnouncementChannel;
 import org.sakaiproject.announcement.api.AnnouncementChannelEdit;
 import org.sakaiproject.announcement.api.AnnouncementMessage;
 import org.sakaiproject.announcement.api.AnnouncementMessageEdit;
-import org.sakaiproject.announcement.api.AnnouncementMessageHeader;
 import org.sakaiproject.announcement.api.AnnouncementMessageHeaderEdit;
 import org.sakaiproject.announcement.cover.AnnouncementService;
 import org.sakaiproject.alias.api.AliasService;
 import org.sakaiproject.alias.api.Alias;
 import org.sakaiproject.announcement.tool.MenuBuilder.ActiveTab;
+import org.sakaiproject.announcement.tool.AnnouncementActionState;
 import org.sakaiproject.authz.api.PermissionsHelper;
 import org.sakaiproject.authz.api.SecurityAdvisor;
 import org.sakaiproject.authz.api.SecurityService;
@@ -98,8 +95,6 @@ import org.sakaiproject.tool.cover.SessionManager;
 import org.sakaiproject.tool.cover.ToolManager;
 import org.sakaiproject.user.api.Preferences;
 import org.sakaiproject.user.api.PreferencesService;
-import org.sakaiproject.user.api.ContextualUserDisplayService;
-import org.sakaiproject.user.api.User;
 import org.sakaiproject.user.api.UserDirectoryService;
 import org.sakaiproject.util.MergedList;
 import org.sakaiproject.util.MergedListEntryProviderBase;
@@ -220,8 +215,7 @@ public class AnnouncementAction extends PagedResourceActionII
    private static final String VIEW_MODE_BYGROUP  = "view.bygroup";
    private static final String VIEW_MODE_MYGROUPS = "view.mygroups";
 
-   // hours * minutes * seconds * milliseconds
-   private static final long MILLISECONDS_IN_DAY = (24 * 60 * 60 * 1000);
+   /** The number of days, by default, before retraction. */
    private static final long FUTURE_DAYS = 7;
    
    private static final String HIDDEN = "hidden";
@@ -464,354 +458,9 @@ public class AnnouncementAction extends PagedResourceActionII
 	}
 
 	/**
-	 * Decorator for the "Message" class. It adds various properties to the decorated real Announcement message.
-	 */
-	static public class AnnouncementWrapper implements AnnouncementMessage
-	{
-		private boolean enforceMaxNumberOfChars;
-
-		private AnnouncementMessage announcementMesssage;
-
-		private boolean editable;
-
-		private String channelDisplayName;
-
-		private int maxNumberOfChars;
-
-		private String range;
-		
-		private String authorDisplayName;
-		
-		public AnnouncementMessage getMessage()
-		{
-			return this.announcementMesssage;
-		}
-
-		/**
-		 * Constructor
-		 * 
-		 * @param message
-		 *        The message to be wrapped.
-		 * @param currentChannel
-		 *        The channel in which the message is contained.
-		 * @param hostingChannel
-		 *        The channel into which the message is being merged.
-		 * @param maxNumberOfChars
-		 *        The maximum number of characters that will be returned by getTrimmedBody().
-		 */
-		public AnnouncementWrapper(AnnouncementMessage message, AnnouncementChannel currentChannel,
-				AnnouncementChannel hostingChannel, AnnouncementActionState.DisplayOptions options, String range)
-		{
-			if (options != null)
-			{
-				this.maxNumberOfChars = options.getNumberOfCharsPerAnnouncement();
-				this.enforceMaxNumberOfChars = options.isEnforceNumberOfCharsPerAnnouncement();
-			}
-			else
-			{
-				// default settings from DisplayOptions class
-				this.maxNumberOfChars = Integer.MAX_VALUE;
-				this.enforceMaxNumberOfChars = false;
-			}
-			this.announcementMesssage = message;
-
-			// This message is editable only if the site matches.
-			this.editable = currentChannel.getReference().equals(hostingChannel.getReference());
-
-			Site site = null;
-
-			try
-			{
-				site = SiteService.getSite(currentChannel.getContext());
-			}
-			catch (IdUnusedException e)
-			{
-				// No site available.
-			}
-
-			if (site != null)
-			{
-				this.channelDisplayName = site.getTitle();
-			}
-			else
-			{
-				this.channelDisplayName = "";
-			}
-
-			// TODO Let's not retrieve the service for each and every message....
-			ContextualUserDisplayService contextualUserDisplayService = (ContextualUserDisplayService) ComponentManager.get("org.sakaiproject.user.api.ContextualUserDisplayService");
-			User author = message.getAnnouncementHeader().getFrom();
-			if ((site != null) && (!this.editable) && (contextualUserDisplayService != null))
-			{
-				this.authorDisplayName = contextualUserDisplayService.getUserDisplayName(author, site.getReference());
-			}
-			if (this.authorDisplayName == null)
-			{
-				this.authorDisplayName = author.getDisplayName();
-			}
-
-			if (range != null)
-			{
-				this.range = range;
-			}
-		}
-
-		/**
-		 * Constructor
-		 * 
-		 * @param announcementWrapper
-		 *        The message to be wrapped.
-		 */
-		public AnnouncementWrapper(AnnouncementWrapper mWrapper)
-		{
-			this.maxNumberOfChars = mWrapper.maxNumberOfChars;
-			this.enforceMaxNumberOfChars = mWrapper.enforceMaxNumberOfChars;
-			this.announcementMesssage = mWrapper.getMessage();
-			
-			this.channelDisplayName = mWrapper.channelDisplayName;
-			this.range = mWrapper.range;
-		}
-
-		/**
-		 * See if the given message was posted in the last N days, where N is the value of the maxDaysInPast parameter.
-		 */
-		private static boolean isMessageWithinLastNDays(AnnouncementMessage message, int maxDaysInPast)
-		{
-			long currentTime = TimeService.newTime().getTime();
-
-			long timeDeltaMSeconds = currentTime - message.getHeader().getDate().getTime();
-
-			long numDays = timeDeltaMSeconds / MILLISECONDS_IN_DAY;
-
-			return (numDays <= maxDaysInPast);
-		}
-
-		/*
-		 * (non-Javadoc)
-		 * 
-		 * @see org.chefproject.core.Message#getHeader()
-		 */
-		public MessageHeader getHeader()
-		{
-			return announcementMesssage.getHeader();
-		}
-
-		/*
-		 * (non-Javadoc)
-		 * 
-		 * @see org.chefproject.core.Message#getBody()
-		 */
-		public String getBody()
-		{
-			return announcementMesssage.getBody();
-		}
-
-		/*
-		 * (non-Javadoc)
-		 * 
-		 * @see org.chefproject.core.Message#getBody()
-		 */
-		public String getTrimmedBody()
-		{
-			if (this.enforceMaxNumberOfChars)
-			{
-				// trim the body, as formatted text
-				String body = announcementMesssage.getBody();
-				StringBuilder buf = new StringBuilder();
-				body = ComponentManager.get(FormattedText.class).escapeHtmlFormattedTextSupressNewlines(body);
-				boolean didTrim = ComponentManager.get(FormattedText.class).trimFormattedText(body, this.maxNumberOfChars, buf);
-				if (didTrim)
-				{
-					if (buf.toString().length() != 0)
-					{
-						buf.append("...");
-					}
-				}
-
-				return buf.toString();
-			}
-			else
-			{
-				return announcementMesssage.getBody();
-			}
-		}
-
-		/*
-		 * (non-Javadoc)
-		 * 
-		 * @see org.chefproject.core.Resource#getUrl()
-		 */
-		public String getUrl()
-		{
-			return announcementMesssage.getUrl();
-		}
-
-		/*
-		 * (non-Javadoc)
-		 * 
-		 * @see org.chefproject.core.Resource#getReference()
-		 */
-		public String getReference()
-		{
-			return announcementMesssage.getReference();
-		}
-
-		/**
-		 * @inheritDoc
-		 */
-		public String getReference(String rootProperty)
-		{
-			return getReference();
-		}
-
-		/**
-		 * @inheritDoc
-		 */
-		public String getUrl(String rootProperty)
-		{
-			return getUrl();
-		}
-
-		/*
-		 * (non-Javadoc)
-		 * 
-		 * @see org.chefproject.core.Resource#getId()
-		 */
-		public String getId()
-		{
-			return announcementMesssage.getId();
-		}
-
-		/*
-		 * (non-Javadoc)
-		 * 
-		 * @see org.chefproject.core.Resource#getProperties()
-		 */
-		public ResourceProperties getProperties()
-		{
-			return announcementMesssage.getProperties();
-		}
-		
-		/**
-		 * returns the range string
-		 * 
-		 * @return
-		 */
-		public String getRange()
-		{
-			return range;
-		}
-
-		/**
-		 * Set the range string
-		 * 
-		 * @return
-		 */
-		public void setRange(String range)
-		{
-			this.range = range;
-		}
-
-		/*
-		 * (non-Javadoc)
-		 * 
-		 * @see org.chefproject.core.Resource#toXml(org.w3c.dom.Document, java.util.Stack)
-		 */
-		public Element toXml(Document doc, Stack stack)
-		{
-			return announcementMesssage.toXml(doc, stack);
-		}
-
-		/*
-		 * (non-Javadoc)
-		 * 
-		 * @see java.lang.Comparable#compareTo(java.lang.Object)
-		 */
-		public int compareTo(Object arg0)
-		{
-			return announcementMesssage.compareTo(arg0);
-		}
-
-		/**
-		 * Returns true if the message is editable.
-		 */
-		public boolean isEditable()
-		{
-			return editable;
-		}
-
-		/**
-		 * Returns the string that is used to show the channel to the user.
-		 */
-		public String getChannelDisplayName()
-		{
-			return channelDisplayName;
-		}
-
-		/*
-		 * (non-Javadoc)
-		 * 
-		 * @see org.chefproject.core.AnnouncementMessage#getAnnouncementHeader()
-		 */
-		public AnnouncementMessageHeader getAnnouncementHeader()
-		{
-			return announcementMesssage.getAnnouncementHeader();
-		}
-		
-		public String getAuthorDisplayName()
-		{
-			return authorDisplayName;
-		}
-
-		/**
-		 * Constructs a list of wrapped/decorated AnnouncementMessages when given a list of unwrapped/undecorated AnnouncementMessages.
-		 * 
-		 * @param messages
-		 *        The list of messages.
-		 * @param currentChannel
-		 *        The current channel being processed.
-		 * @param hostingChannel
-		 *        The default channel of the page into which this list is being merged.
-		 * @param maxNumberOfDaysInThePast
-		 *        Messages over this limit will not be included in the list.
-		 * @param maxCharsPerAnnouncement
-		 *        The maximum number of characters that will be returned when getTrimmedBody() is called.
-		 */
-		static private List<AnnouncementWrapper> wrapList(List<AnnouncementMessage> messages, AnnouncementChannel currentChannel, AnnouncementChannel hostingChannel,
-				AnnouncementActionState.DisplayOptions options)
-		{
-			// 365 is the default in DisplayOptions
-			int maxNumberOfDaysInThePast = (options != null) ? options.getNumberOfDaysInThePast() : 365;
-			 
-
-			List<AnnouncementWrapper> messageList = new ArrayList<>();
-
-			Iterator<AnnouncementMessage> it = messages.iterator();
-
-			while (it.hasNext())
-			{
-				AnnouncementMessage message = it.next();
-
-				// See if the message falls within the filter window.
-				// note: the default of enforceNumberOfDaysInThePastLimit is false
-				if (options != null && options.isEnforceNumberOfDaysInThePastLimit() && !isMessageWithinLastNDays(message, maxNumberOfDaysInThePast))
-				{
-					continue;
-				}
-
-				messageList.add(new AnnouncementWrapper(message, currentChannel, hostingChannel, options,
-						getAnnouncementRange(message)));
-			}
-
-			return messageList;
-		}
-
-	}
-
-	/**
 	 * get announcement range information
 	 */
-	private static String getAnnouncementRange(AnnouncementMessage a)
+	static String getAnnouncementRange(AnnouncementMessage a)
 	{
 		if (a.getProperties().getProperty(ResourceProperties.PROP_PUBVIEW) != null
 				&& a.getProperties().getProperty(ResourceProperties.PROP_PUBVIEW).equals(Boolean.TRUE.toString()))
@@ -1090,7 +739,7 @@ public class AnnouncementAction extends PagedResourceActionII
 							// If any message is allowed to be removed
 							// Also check to see if the AnnouncementWrapper object thinks
 							// that this message is editable from the default site.
-							if (message.editable && channel.allowRemoveMessage(message))
+							if (message.isEditable() && channel.allowRemoveMessage(message))
 							{
 								menu_delete = true;
 								break;
@@ -2149,13 +1798,12 @@ public class AnnouncementAction extends PagedResourceActionII
 			else
 				context.put("pubview", Boolean.FALSE);
 
-			// Set inital release date to today
+			// TODO: Track any usage of these Time objects and convert to java.time
+			// Set initial release date to today
 			final Time currentTime = TimeService.newTime();
 			context.put(AnnouncementService.RELEASE_DATE, currentTime);
 			
-			// Set inital retract date to 60 days from now
-			final long futureTimeLong = currentTime.getTime() + MILLISECONDS_IN_DAY * FUTURE_DAYS;			
-			final Time futureTime = TimeService.newTime(futureTimeLong);
+			final Time futureTime = defaultRetractTime();
 
 			context.put(AnnouncementService.RETRACT_DATE, futureTime);
 			
@@ -2210,15 +1858,13 @@ public class AnnouncementAction extends PagedResourceActionII
 			try 
 			{
 				retractDate = edit.getProperties().getTimeProperty(AnnouncementService.RETRACT_DATE);
-				
+
 				context.put("useRetractDate", Boolean.valueOf(true));
 				specify = true;
 			} 
 			catch (Exception e) 
 			{
-				// Set inital retract date to approx 2 months from today
-				final long futureTimeLong = TimeService.newTime().getTime() + MILLISECONDS_IN_DAY * FUTURE_DAYS;			
-				retractDate = TimeService.newTime(futureTimeLong);
+				retractDate = defaultRetractTime();
 			}
 
 			context.put(AnnouncementService.RETRACT_DATE, retractDate);
@@ -2285,17 +1931,14 @@ public class AnnouncementAction extends PagedResourceActionII
 				}
 				else
 				{
-					// Set inital retract date to 60 days from now				
-				final long futureTimeLong = TimeService.newTime().getTime() + MILLISECONDS_IN_DAY * FUTURE_DAYS;			
-				retractDate = TimeService.newTime(futureTimeLong);
+				retractDate = defaultRetractTime();
+
 				context.put("useRetractDate", Boolean.valueOf(false));
 				}
 			} 
 			catch (Exception e) 
 			{
-				// Set inital retract date to approx 2 months from today
-				final long futureTimeLong = TimeService.newTime().getTime() + MILLISECONDS_IN_DAY * FUTURE_DAYS;			
-				retractDate = TimeService.newTime(futureTimeLong);
+				retractDate = defaultRetractTime();
 			}
 
 			context.put(AnnouncementService.RETRACT_DATE, retractDate);
@@ -2342,6 +1985,27 @@ public class AnnouncementAction extends PagedResourceActionII
 		return template + "-revise";
 
 	} // buildReviseAnnouncementContext
+
+	/**
+	 * Calculate the default retract date from now.
+	 *
+	 * The duration in days is set in {@link #FUTURE_DAYS}.
+	 * @deprecated Migrate away from Time and use @link{defaultRetractDate}
+	 * @return a Time in the future when the message will be retracted.
+	 **/
+	@Deprecated
+	private Time defaultRetractTime() {
+		return TimeService.newTime(defaultRetractDate().toEpochMilli());
+	}
+
+	/** Calculate the default retract date from now.
+	 *
+	 * The duration in days is set in {@link #FUTURE_DAYS}.
+	 * @return an instant in the future when the message will be retracted.
+	 **/
+	private Instant defaultRetractDate() {
+		return Instant.now().plus(Duration.ofDays(FUTURE_DAYS));
+	}
 
 	/**
 	 * Build the context for viewing announcement content
